+++ src/trackers/classes/crawl.js
@@ -3,6 +3,9 @@ const {median, std} = require('mathjs')
 const shared = require('./../helpers/sharedData.js')
 const CommonRequest = require('./commonRequest.js')
 const sharedData = require('./../helpers/sharedData.js')
+const {getFingerprintWeights} = require('./../helpers/fingerprints.js')
+
+const API_COUNT_THRESHOLD = 15
 
 class Crawl {
     constructor () {
@@ -106,7 +109,7 @@ function _processSite (crawl, site) {
         const apisUsed = Object.keys(apis)
         
         let tracking = false
-        if (apisUsed.length >= 15) {
+        if (apisUsed.length >= API_COUNT_THRESHOLD) {
             tracking = true
         }
         
@@ -164,37 +167,6 @@ function _getDomainSummaries (crawl) {
     return domainSummary
 }
 
-function _getFingerprintWeights (crawl) {
-    const apiWeights = {}
-    let maxWeight = 0
-    const setMaxWeight = []
-
-    for(const [api, apiScores] of Object.entries(crawl.fpWeights.apis)) {
-        const trackingWt = apiScores.tracking / crawl.fpWeights.scripts.tracking
-        const nontrackingWt = apiScores.nontracking / crawl.fpWeights.scripts.nontracking
-
-        // api is always used for tracking, set to highest weight we see
-        if (nontrackingWt === 0) {
-            setMaxWeight.push(api)
-        } else {
-            const weight = trackingWt / nontrackingWt
-            apiWeights[api] = weight
-
-            if (weight > maxWeight) {
-                maxWeight = weight
-            }
-        }
-    }
-
-    if (setMaxWeight.length) {
-        setMaxWeight.forEach(api => {
-            apiWeights[api] = maxWeight
-        })
-    }
-
-    return apiWeights
-}
-
 function _getEntitySummaries (crawl) {
     delete crawl.entityPrevalence.undefined
 
@@ -236,7 +208,7 @@ function _writeSummaries (crawl) {
     csv = csv.sort((a, b) => b[1] - a[1])
     fs.writeFileSync(`${shared.config.trackerDataLoc}/build-data/generated/entity_prevalence.csv`, csv.reduce((str, row) => {str += `"${row[0]}",${row[1]},${row[2]},${row[3]}\n`; return str}, 'Entity,Total Prevalence,Tracking Prevalence,Non-tracking Prevalence\n'))
 
-    fs.writeFileSync(`${shared.config.trackerDataLoc}/build-data/generated/api_fingerprint_weights.json`, JSON.stringify(_getFingerprintWeights(crawl), null, 4))
+    fs.writeFileSync(`${shared.config.trackerDataLoc}/build-data/generated/api_fingerprint_weights.json`, JSON.stringify(getFingerprintWeights(crawl), null, 4))
 }
 
 function updateEntityPrevalence (crawl) {
